refactor: generic thermostat tests step 1 - splitting files
