+++ tests/components/generic_thermostat/const.py
@@ -1,0 +1,13 @@
+"""Constants for Generic Thermostat tests."""
+ENTITY = "climate.test"
+ENT_SENSOR = "sensor.test"
+ENT_SWITCH = "switch.test"
+HEAT_ENTITY = "climate.test_heat"
+COOL_ENTITY = "climate.test_cool"
+ATTR_AWAY_MODE = "away_mode"
+MIN_TEMP = 3.0
+MAX_TEMP = 65.0
+TARGET_TEMP = 42.0
+COLD_TOLERANCE = 0.5
+HOT_TOLERANCE = 0.5
+TARGET_TEMP_STEP = 0.5
