+++ tests/components/generic_thermostat/shared.py
@@ -1,0 +1,27 @@
+"""Shared functions for the generic_thermostat component test."""
+from homeassistant import core as ha
+from homeassistant.const import SERVICE_TURN_OFF, SERVICE_TURN_ON, STATE_OFF, STATE_ON
+from homeassistant.core import callback
+
+from tests.components.generic_thermostat.const import ENT_SENSOR, ENT_SWITCH
+
+
+def _setup_sensor(hass, temp):
+    """Set up the test sensor."""
+    hass.states.async_set(ENT_SENSOR, temp)
+
+
+def _setup_switch(hass, is_on):
+    """Set up the test switch."""
+    hass.states.async_set(ENT_SWITCH, STATE_ON if is_on else STATE_OFF)
+    calls = []
+
+    @callback
+    def log_call(call):
+        """Log service calls."""
+        calls.append(call)
+
+    hass.services.async_register(ha.DOMAIN, SERVICE_TURN_ON, log_call)
+    hass.services.async_register(ha.DOMAIN, SERVICE_TURN_OFF, log_call)
+
+    return calls
